+++ tests/test_pdp/test_pdp_with_sql.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import sessionmaker, scoped_session
 from py_abac.pdp import PDP, EvaluationAlgorithm
 from py_abac.policy import Policy
 from py_abac.provider.base import AttributeProvider
-from py_abac.request import Request
+from py_abac.request import AccessRequest
 from py_abac.storage.sql import SQLStorage
 from py_abac.storage.sql.model import Base
 from ..test_storage.test_sql import create_test_sql_engine
@@ -407,7 +407,7 @@ def st(session):
 ])
 def test_is_allowed_deny_overrides(st, desc, request_json, should_be_allowed):
     pdp = PDP(st, EvaluationAlgorithm.DENY_OVERRIDES, [EmailsAttributeProvider()])
-    request = Request.from_json(request_json)
+    request = AccessRequest.from_json(request_json)
     assert should_be_allowed == pdp.is_allowed(request)
 
 
@@ -646,7 +646,7 @@ def test_is_allowed_deny_overrides(st, desc, request_json, should_be_allowed):
 ])
 def test_is_allowed_allow_overrides(st, desc, request_json, should_be_allowed):
     pdp = PDP(st, EvaluationAlgorithm.ALLOW_OVERRIDES, [EmailsAttributeProvider()])
-    request = Request.from_json(request_json)
+    request = AccessRequest.from_json(request_json)
     assert should_be_allowed == pdp.is_allowed(request)
 
 
@@ -885,7 +885,7 @@ def test_is_allowed_allow_overrides(st, desc, request_json, should_be_allowed):
 ])
 def test_is_allowed_highest_priority(st, desc, request_json, should_be_allowed):
     pdp = PDP(st, EvaluationAlgorithm.HIGHEST_PRIORITY, [EmailsAttributeProvider()])
-    request = Request.from_json(request_json)
+    request = AccessRequest.from_json(request_json)
     assert should_be_allowed == pdp.is_allowed(request)