+++ tests/test_policy/test_conditions/test_object.py
@@ -8,7 +8,7 @@ from marshmallow import ValidationError
 from py_abac.context import EvaluationContext
 from py_abac.policy.conditions.object import EqualsObject
 from py_abac.policy.conditions.schema import ConditionSchema
-from py_abac.request import Request
+from py_abac.request import AccessRequest
 
 
 class TestCollectionCondition(object):
@@ -48,7 +48,7 @@ class TestCollectionCondition(object):
          False),
     ])
     def test_is_satisfied(self, condition, what, result):
-        request = Request(subject={"attributes": {"what": what}}, resource={}, action={}, context={})
+        request = AccessRequest(subject={"attributes": {"what": what}}, resource={}, action={}, context={})
         ctx = EvaluationContext(request)
         ctx.ace = "subject"
         ctx.attribute_path = "$.what"